important hotfix
Very simple fix for follow-up attacks, kills, hits, and relic triggers.

Previously "2 kills" would mean two kills EVERY turn. This was unintended behavior and is now fixed.

Now they are front-loaded, however, meaning that it is assumed they all happen in one turn which can cause some issues in the future with really fast rotations. I will definitely look more into this.

This is intended to be only a quick fix to an important issue.